+++ packages/guide/src/app/journal/components/produce/produce.component.ts
@@ -1,4 +1,4 @@
-import { Component, computed } from '@angular/core';
+import { Component, computed, inject } from '@angular/core';
 import { BaseJournalPageComponent } from '../base-journal-page/base-journal-page.component';
 import { BaseCrop, Crop, FruitPlant, FruitTree, Item, MinimalItem, Season, Seasons } from '@ci/data-types';
 import { combineLatest, of, switchMap } from 'rxjs';
@@ -6,13 +6,14 @@ import { FormControl, FormGroup } from "@angular/forms";
 import { FilterForm } from "../../../shared/types/filter-form.type";
 import { nonNullable } from "@ci/util";
 import { ToDoContext } from "../../../core/types/to-do-context.type";
+import { SettingsService } from "../../../shared/services/settings.service";
 
 @Component({
     selector: 'app-produce',
     templateUrl: './produce.component.html',
 })
 export class ProduceComponent extends BaseJournalPageComponent<MinimalItem | Crop | FruitPlant | FruitTree> {
-
+    isBeta = inject(SettingsService).getSettings().useBeta
     toDoContext = computed<ToDoContext | undefined>(() => {
 
         const selectedTabIndex = this.selectedTabIndex()
@@ -23,7 +24,9 @@ export class ProduceComponent extends BaseJournalPageComponent<MinimalItem | Cro
                 ? "journal_animal_produce"
                 : selectedTabIndex === 2
                     ? "artisan"
-                    : undefined;
+                    : selectedTabIndex === 3
+                        ? "journal_ocean_produce"
+                        : undefined;
     });
 
     constructor() {
@@ -69,7 +72,21 @@ export class ProduceComponent extends BaseJournalPageComponent<MinimalItem | Cro
                     2
                 )
             },
+
         ];
+
+        if (this.isBeta) {
+            this.tabs.push(
+                {
+                    title: 'Ocean',
+                    data: this.getFilteredJournalData(
+                        this._database.fetchJournalOrder$('journal-ocean-products'),
+                        this._database.fetchItems$(),
+                        3
+                    )
+                }
+            )
+        }
 
         this.activateTabFromRoute(this.tabs.map(tab => tab.title));
     }