+++ packages/pak-files-parser/src/main.ts
@@ -246,6 +246,7 @@ AvailableLanguages.forEach(lang => {
 
 
             'journal-animal-products': new JournalOrderDbGenerator('Produce/DT_JournalAnimalProducts.json'),
+            'journal-ocean-products': new JournalOrderDbGenerator('Produce/DT_JournalOcean.json'),
             'journal-artisan-products': new JournalOrderDbGenerator('Produce/DT_JournalArtisanProducts.json'),
             'journal-crops': new JournalOrderDbGenerator('Produce/DT_JournalCrops.json'),